Backend-Azure-Functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions==1.17.0
azure-storage-blob==12.17.0
Pillow==9.5.0